requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
customtkinter==5.2.2
darkdetect==0.8.0
discord.py==2.7.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
packaging==26.0
pillow==12.1.1
propcache==0.4.1
python-dotenv==1.2.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
